Coordinate System Fixes
I think the coordinate systems should all be in the pure field coordinates now. It is highly likely there are errors since I do not remember all the sign conventions of the values.

Also, please do not use the "crude" method as I haven't thought of a good way to fix it yet.